General Health Query Chatbot/requirements.txt
Flask==2.3.3
google-generativeai==0.5.4
requests==2.31.0